# backend/swagger.yaml
openapi: 3.0.0
info:
  title: ECC Ministry Application API
  description: 'API for Coding Challenge (IS24 - Full Stack Developer at Ministry of Education & Child Care)'
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api'
    description: Server for development
tags: 
  - name: ECC Ministry Products
    description: All products maintained or being worked on by the Ministry of Education and Child Care
paths:
  /health:
    get:
      tags: 
        - ECC Ministry Products 
      summary: Check health of the server
      responses:
        '200': 
          description: A successful health check returns "OK"
          content:
            application/json:
              schema: 
                type: array
  /products:
    get:
      tags: 
        - ECC Ministry Products
      summary: Returns a list of products.
      responses:
        '200':    
          description: A JSON array of products
          content:
            application/json:
              schema: 
                type: array
    post:
      tags: 
        - ECC Ministry Products
      summary: Create product
      requestBody: 
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/Product'
            examples:
              Lisa:
                value: 
                  productName: "Sample Product Name One"
                  productOwnerName: "Dee Sember"
                  developers: ["Robin Banks", "Anita Bath"]
                  scrumMasterName: "Sue Flay"
                  methodology: "Agile"
                  startDate: "2023/10/13"
              Alan:
                value: 
                  productName: "Sample Product Name Two"
                  productOwnerName: "Bill Ding"
                  developers: ["Robin Banks", "Anita Bath", Paige Turner]
                  scrumMasterName: "Barry Cade"
                  methodology: "Waterfall"
                  location: "github.com/bcgov/24"
      responses: 
        '201':    
          description: Product successfully created
        '400':
          description: Product could not be created
  /products/{id}:
    get:
      tags: 
        - ECC Ministry Products
      summary: Find product by ID
      parameters: 
        - name: id
          in: path
          description: ID of product that needs to be fetched
          required: true 
          schema:
            type: integer
            format: int64
      responses:
        '200':    
          description: Product with given ID was found
        '404': 
          description: Product with given ID was not found
    put:
      tags: 
        - ECC Ministry Products
      summary: Update product information
      parameters:
        - name: id 
          in: path
          description: Product ID that you want to edit information
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Update an existing product in the table
        content: 
          application/json:  
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product successfully update
        '404': 
          description: Product could not be updated

components:
  schemas:
    Product:
      type: object
      properties:
        productName:
          type: string
          example: Test Product Name
        productOwnerName:
          type: string
          example: Test Product Owner
        developers: 
          type: array
          example: [dev1, dev2, dev3, dev4, dev]
        scrumMasterName:
          type: string 
          example: Test Scrum Master Name
        methodology: 
          type: string
          example: Agile
        startDate:
          type: string
          example: 2023/09/05
        location:
          type: string
          example: 'github/bcgov/test-name'
      required:
        - productName
        - productOwnerName
        - developers
        - scrumMasterName
        - methodology

          